requirements.txt
torch>=1.12
torchvision
numpy
scikit-learn
matplotlib
pillow
tqdm
opencv-python
ortools
seaborn
